Added new scheduling functionality to benchmark algorithm

+++ assets/bench_mark.py
@@ -11,36 +11,68 @@ import platform
 import subprocess
 import threading
 import socket
+import datetime
+import clientwindow
 log = logging.getLogger('bench_mark.py')
 resArray = []
 string = ""
 host = ["10.41.31.99", "google.com"]
 socketIO = None
-bench_thread = None
+bench_thread = []
+threadNames = []
+terminated = False
 
-def init(*args, **kw):
-    global bench_thread
+def init(times,socket):
+    global bench_thread,terminated
     global socketIO
-    socketIO = args[0]
-    stop()
-    start(10)
+    socketIO = socket
+    if len(threadNames) is not 0:
+        stop(False)
+    display_str = "Benchmark Execution Scheduled at "
+    for i in range(len(times)):
+        timeNow =  datetime.datetime.today()
+        display_str = display_str + times[i] + " "
+        schedTime = datetime.datetime.strptime(times[i],"%H:%M")
+        delta = schedTime - timeNow
+        start(delta.seconds + 1,"bench_thread"+str(i),schedTime)
+    logger.print_on_console(display_str)
 
-def start(t=1800):
-    global bench_thread
-    bench_thread = threading.Timer(t, execute)
+
+def start(t,name,schedTime):
+    global bench_thread,threadNames,terminated
+    terminated = False
+    bench_thread = threading.Timer(t, execute,[name,schedTime])
+    bench_thread.setName(name)
+    threadNames.append(bench_thread)
     bench_thread.start()
 
-def stop():
-    if (bench_thread is not  None) and bench_thread.isAlive():
-        bench_thread.cancel()
+def stop(terminateExec):
+    global terminated,bench_thread,threadNames
+    bench_thread = []
+    delete_memory_test_file()
+    terminated = True
+    if terminateExec:
+        return
+    for i in range(len(threadNames)):
+        threadNames[i].cancel()
+        time.sleep(1)
+    threadNames = []
+    return
+   
+def delete_memory_test_file():
+    if os.path.exists('output.txt'):
+        os.unlink('output.txt')
 
 def cpu():
+    global terminated
     logger.print_on_console("Running CPU benchmark")
     for j in range(3):
         beforeCpu = time.time()
         arr = []
         global string
-        for i in range(0,300):
+        for i in range(0,350):
+            if terminated:
+                return -99
             x = random.randint(1,100001)
             arr.append(pow(pow(x, random.randint(1,10)),1/random.randint(1,10))) 
             random.shuffle(arr)
@@ -59,34 +91,46 @@ def cpu():
     return sum/len(resArray)
         
 def memory():
-    logger.print_on_console("Running memory benchmark")
-    global string
     try:
-        os.unlink("output.txt")
+        global terminated
+        logger.print_on_console("Running memory benchmark")
+        global string
+        delete_memory_test_file()
+        for x in range(4):
+            string = string + string
+        for j in range(2): 
+            if terminated:
+                return -99
+            beforeFile = time.time()
+            if os.path.exists("output.txt"):
+                outputFile = os.stat("output.txt")
+                if outputFile.st_size/100000 > 4000:
+                    os.unlink("output.txt")
+            for i in range(40):
+                if terminated:
+                    return -99
+                f = open("output.txt", "a")
+                f.write(string)
+                f.close()
+            for i in range(4):
+                if terminated:
+                    return -99
+                f = open("output.txt", "r")
+                f.read()
+                f.close()
+            os.unlink("output.txt")
+            afterFile = time.time()
+            resArray.append(afterFile - beforeFile)
+        sum = 0
+        for i in range(len(resArray)):
+            sum += resArray[i]
+            
+        result = sum/len(resArray)
+        print("Memory Performance",result)
+        return result
     except Exception as e:
-        log.error(e)
-    for x in range(4):
-        string = string + string
-    for j in range(2): 
-        beforeFile = time.time()
-        for i in range(40):
-            f = open("output.txt", "a")
-            f.write(string)
-            f.close()
-        for i in range(4):
-            f = open("output.txt", "r")
-            f.read()
-            f.close()
-        os.unlink("output.txt")
-        afterFile = time.time()
-        resArray.append(afterFile - beforeFile)
-    sum = 0
-    for i in range(len(resArray)):
-        sum += resArray[i]
-        
-    result = sum/len(resArray)
-    print("Memory Performance",result)
-    return result
+        logger.print_on_console("Error in Benchmark Execution")
+        return -99
 
 def selectionSort(arr):
     for i in range(len(arr)): 
@@ -106,7 +150,7 @@ def convertToBinary(rr):
 		rr[i] = dec2bin(rr[i])
   
 def ping(host):
-	p = subprocess.Popen(["ping.exe",host], stdout = subprocess.PIPE)
+	p = subprocess.Popen(["ping.exe",host], stdout = subprocess.PIPE, shell = True)
 	result = p.communicate()[0]
 	for row in str(result).split("\\r\\n"):
 		if 'Average' in row:
@@ -116,14 +160,18 @@ def ping(host):
 
 def network():
     logger.print_on_console("Running network benchmark")
-    global host
+    global host,terminated
+    if terminated:
+        return -99,-99
     loss = 0
     sent = 0
     for i in range(2):
         for j in range(len(host)):
+            if terminated: 
+                return -99,-99
             sent += 1
             p = ping(host[j])
-            if p == -1:
+            if p == -1: 
                 loss += 1
             else:
                 resArray.append(ping(host[j]))
@@ -135,18 +183,36 @@ def network():
     logger.print_on_console("Packet Lost ",loss)
     return sum/len(resArray), ((sent - loss)/sent)
 
-def execute():
-    global string,resArray
-    string = ""
-    resArray = []
-    logger.print_on_console("Starting Benchmark execution")
-    logger.print_on_console("Executing system benchmark")
-    a,percent = network()
-    b = cpu()
-    c = memory()
-    #a = b = c = 452
-    result = (a+b+c)/3
-    logger.print_on_console("Benchmark result: ",result)
-    socketIO.emit('benchmark_ping',{'cpuscore':b,"time":time.ctime(),"memoryscore":c,"networkscore":a,"percent_received":str(percent * 100) + '%',"hostip":socket.gethostbyname(socket.gethostname()),"hostname":os.environ['username'],"systemscore":result})
-    start(300)
+def execute(name,schedTime):
+    global string,resArray,terminated,socketIO
+    timeNow =  datetime.datetime.today()
+    delta = schedTime - timeNow
+    start(delta.seconds,name,schedTime)
+    if clientwindow.execution_flag:
+        logger.print_on_console("Execution in progress, t erminating Benchmark Execution")
+        return
+    if terminated:
+        return
+    try:
+        string = ""
+        resArray = []
+        logger.print_on_console("Starting Benchmark execution")
+        a,percent = network()
+        b = cpu()
+        c = memory()
+        if a == -99 or b == -99 or c == -99:
+            logger.print_on_console("Terminating Benchmark Execution")
+            return
+        result = (a+b+c)/3
+        logger.print_on_console("Benchmark result: ",result)
+        if terminated:
+            logger.print_on_console("Terminating Benchmark Execution")
+            return 
+        socketIO.emit('benchmark_ping',{'cpuscore':b,"memoryscore":c,"networkscore":a,"percent_received":str(percent * 100) + '%',"hostip":socket.gethostbyname(socket.gethostname()),"hostname":os.environ['username'],"systemscore":result,"time":str(datetime.datetime.isoformat(datetime.datetime.now()))})
+        return
+    except Exception as e:
+        delete_memory_test_file()
+        logger.print_on_console("Error in Benchmark Execution, terminating.")
+        log.error(e)
+        return
